# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank system
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
...
  /payments:
    post:
      tags:
      - Payment Management
      summary: Transfer money between cards
      description: Allows a user to transfer money between their own cards
      operationId: transferMoney
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentDto"
        required: true
      responses:
        "200":
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "400":
          description: Invalid input data or insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                Validation Error:
                  description: Validation Error
                  value:
                    timestamp: 2024-07-30T12:34:56.789Z
                    message: Validation failed
                    code: 400
                    validationErrors:
                    - field: amount
                      message: Amount must not be less than zero
                Insufficient Funds:
                  description: Insufficient Funds
                  value:
                    timestamp: 2024-07-30T12:34:56.789Z
                    message: There are not enough funds on the card
                    code: 400
        "403":
          description: Card is blocked or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 is blocked for operation"
                code: 403
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found for user 1"
                code: 404
...
  /cards:
    get:
      tags:
      - User Card Management
      summary: Get all user cards
      description: Retrieves a paginated list of all cards belonging to the authenticated
        user
      operationId: getUserCards
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 6
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "balance,ASC"
          items:
            type: string
      responses:
        "200":
          description: Cards found
          content:
            application/json: {}
  /cards/{id}:
    get:
      tags:
      - User Card Management
      summary: Get card by ID
      description: "Retrieves a specific card by its ID, if it belongs to the authenticated\
        \ user"
      operationId: getCardById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Card found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found for user 1"
                code: 404
...
  /admin/cards/{id}:
    get:
      tags:
      - Admin Card Management
      summary: Get card by ID
      description: Retrieves card details by its ID
      operationId: getCardById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Card found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardDto"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found"
                code: 404
    delete:
      tags:
      - Admin Card Management
      summary: Delete card by ID
      description: Deletes a card by its ID
      operationId: deleteCardById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Card deleted successfully
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2024-07-30T12:34:56.789Z
                message: "Card by id: 1 not found"
                code: 404
  /admin/cards/user/{id}:
    get:
      tags:
      - Admin Card Management
      summary: Get all cards for a user
      description: Retrieves a paginated list of all cards for a specified user
      operationId: getUserCards_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 6
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          default:
          - "validityPeriod,ASC"
          items:
            type: string
      responses:
        "200":
          description: Cards found
          content:
            application/json: {}
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        timestamp:
          type: string
          format: date-time
          description: Timestamp of the error
        message:
          type: string
          description: Error message
          example: Validation failed
        code:
          type: integer
          format: int32
          description: HTTP status code
          example: 400
        validationErrors:
          type: array
          description: List of validation errors
          items:
            $ref: "#/components/schemas/ValidationError"
    ValidationError:
      type: object
      description: Details of a validation error
      properties:
        field:
          type: string
          description: Field where the error occurred
          example: username
        message:
          type: string
          description: Error message for the field
          example: Username must not be empty
...
    PaymentDto:
      type: object
      description: Request for a payment transfer
      properties:
        fromCardId:
          type: integer
          format: int64
          description: ID of the card to transfer from
          example: 1
        toCardId:
          type: integer
          format: int64
          description: ID of the card to transfer to
          example: 2
        amount:
          type: number
          description: Amount to transfer
          example: 100.5
        comment:
          type: string
          description: Comment for the payment
          example: Payment for goods
      required:
      - fromCardId
      - toCardId
    CardDto:
      type: object
      description: Card details
      properties:
        cardId:
          type: integer
          format: int64
          description: Card ID
          example: 1
        cardNumber:
          type: string
          description: Masked card number
          example: '**** **** **** 9010'
        validityPeriod:
          type: string
          format: date
          description: Expiration date of the card
          example: 2028-12-31
        statusCard:
          type: string
          description: Current status of the card
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
          - PENDING_BLOCKED
        balance:
          type: number
          description: Current balance of the card
          example: 1000.0
...
  securitySchemes:
    bearerAuth:
      type: http
      name: bearerAuth
      scheme: bearer
      bearerFormat: JWT